Introduce processing scripts
This patch introduces the notion of a "processing script," in which the
HTML processing steps for an endpoint are defined via a declarative
interface.  Processing scripts are defined in YAML files under the new
processing/ directory, loaded at startup time, and held in memory as part
of the application object.

In this introductory commit, only the VE markup stripping steps are
incorporated into the new processing scripts, but this structure could
(and, in future commits, should) be used to define all processing steps.

The goals of this change include:

* To make the processing steps involved for each endpoint easier to see
  at a glance;

* To simplify the processing code away from a tangle of special-purpose
  library methods and toward a consistent, predictable pipeline;

* To improve our ability to refine the particular processing steps being
  performed on a per-endpoint basis;

* To avoid changes intended for one endpoint causing bugs in another (and
  to remove the developer's need to worry about same);

* To make the project more useful to current and potential third party
  users (such as Kiwix) by improving its configurability;

* To provide a basis for further abstracting units of "processing work" so
  that we can more easily control the amount of CPU-intensive work
  performed per tick in the node event loop.

Change-Id: If65215d47818a32a256ae641f9714a147c3e8eb5