mobile-sections: fix issue with missing section text
In some cases there is a heading inside the lead section which
comes from a transclusion, so Parsoid doesn't generate a new section
for it. When that happened before this patch it erased the
earlier collected section text. Now we're adding a new condition to
check if there is anything coming before this to avoid this situation.

Example:
http://0.0.0.0:6927/pt.wikipedia.org/v1/page/mobile-sections/Hipotiroidismo/50805952

Bumped patch version of mobile-sections.

Bug: T209158
Change-Id: I8d43e8f033446bc0652793da3aaca54cf37bae86