Hygiene: Remove 'formatted' endpoints and supporting code
The /page/formatted and /page/formatted-lead endpoints were an early
prototype of next-generation page loading. They were abandoned in favor
of the current PCS plans, but for some time the supporting code was still
relied upon in the /page/summary endpoint for generating extracts. This
is no longer the case, so we can now drop this code and greatly simplify
the processing for page JSON requests.

Change-Id: Ie14c7e8df21bac5f140d7659f7f95059e2bbfdc2