Media: Reduce min image size
Per Cooltey's report on #wikimedia-mobile this morning, the media endpoint
misses the page image (i.e., the Android lead image) for :en:Malaysia.
This is because its on-page height is 63px and our minimum for filtering
is 64.

This is probably not the only case with an SVG flag icon as a page image.
Let's reduce the minimum size, since the size filtering is mostly a backup
for filtering by class (such as .noviewer).

Change-Id: I1e4c5d2e6c884d6520e1f36db5df737868687ed7