Dev: Stop using mocha-eslint
As discovered by Krinkle[1], mocha-eslint pulls in an older version of
eslint (^4.2.0, currently resolving to 4.19.1) than is required by eslint-
config-wikimedia (^5.9.0).  This is true even for the current mocha-eslint
version.

This patch bumps the requested eslint version from ^4.19.1 to ^5.9.0
and drops the mocha-eslint dependency in favor of a direct eslint
invocation in the npm test script. One minor spacing change was needed to
comply with the update.

[1] https://gerrit.wikimedia.org/r/#/c/mediawiki/services/mobileapps/+/451399/

Change-Id: I4081e8e4b5ea9d7acd53a72f0f443f8ffdf72142