Summary: Get modified date with regexes to avoid unneeded Document parse
In the event that we should return empty extracts, we're still currently
parsing the entire page into a domino Document just to get the modified
date. This adds a getModifiedFromHtml function that uses regexes to get
the modified date from the raw Parsoid HTML, allowing us to avoid a full
Document parse where not needed for extracts.

Tests are added for both the old getModified and new getModifiedFromHtml
functions.

Change-Id: Ia00f93e0095a9b3578b19e3998fe2c33dcb0dd73