Set timeout for backend /page/html requests
Our p99 latencies for the summary endpoint, both internal and external,
currently spike to over 4 minutes.  I suspect this is related to the fact
that we're currently waiting for requests to time out twice, for two
minutes each time, for large pages due to T184498.  This patch sets a one
minute timeout to test this theory and to reduce four-minute delays to
two-minute delays in this situation.

Note: This timeout value will also need to be added to the config template
in mobileapps/deploy and deployed to have any effect in production.

Change-Id: I7954891e68f8c99f4edd2ed8ea13200afc029dc4