Dev: add css and js endpoints to diff tests
Added a new optional parameter to the TestSpec constructor, allowing
to pass in an options object with a suffix string.

Added the pagelib js, the pagelib/base/site CSS.

Added the pagelib JS file to .eslintignore since it's missing the
required newline at the end of the file and I don't want to alter
the contents of the output.

Change-Id: Iaf0995924160172fb9d73b6dea5cd40d922f9efe