Dev: added example of static file for prototyping
This is a very simple way to quickly prototype something using static
files. Just add files to the static folder or a subdirectory and have it
served.

This would be only useful for servers that are not behind RESTBase.
For prototyping a server like our labs instance might be good enough.
Example: http://appservice.wmflabs.org/static/proto/example1.json
The above URL should be available after this patch is merged.

Change-Id: I3ed94fbe4b61a1fdafb01f77fa5e57d4c95cb65d