CSS hotfix for Android
1) Android bundles the pagelib CSS, so it would take a while until
Android app users would get the fix in [1].

2) Had to modify the CSS in the PR to avoid a short flicker of a white
bar while loading a page in a non-white theme since it took a bit to set
the theme on the root html element.

3) This brings back the base styles, which had caused issues with dark
mode in the past, but should fix the heading. Combined with the pagelib
change this should be a good temporary fix.

4) It also changes the order of ResourceLoader modules used.
Order does make a difference in CSS. At least we want
skins.minerva.base.styles to come first. Not sure about the other ones.

5) Also needed to unset the height property on the html and body tags
to avoid the native footer on Android overlapping the content at the
bottom of the page.

6) Updated expected diff test result for the base CSS endpoint.

7) Longer-term we should follow JonR's advice and roll our own styles.

8) I recommend the Android app switch to live pagelib CSS[2] and using
meta.wikimedia.org as the host name for base CSS and pagelib CSS/JS.

[1] https://github.com/wikimedia/wikimedia-page-library/pull/171
[2] https://meta.wikimedia.org/api/rest_v1/data/css/mobile/pagelib

Bug: T214728
Bug: T214714
Change-Id: Iddd4646263531ec9da28acdf0b0448035fcdd056