Pull in styles previously obtained through ResourceLoader
This pulls in local copies of the files defining the styles which the
service previously requested from ResourceLoader. Files are copied over
mostly directly, with annotations at the top indicating their original
source and the latest ID to have touched the file at the time of copying.

Note that @import directives have been altered to reflect local filenames
and directory structure, and unnecessary @import statements have been
removed.

These will be compiled with lessc and the base CSS endpoint will be
updated to use them in the following patch.

Bug: T220045
Change-Id: I532f34c92fa547469f5db96a41aaffa600d633cf