Update /data/css/mobile/base to use styles compiled from static files
Incorporates the "less" module to compile and minify the CSS/Less files
now included in the repo into a minified stylesheet to be served via
/data/css/mobile/base. Updates lib/css.js and related code accordingly.

Note that the Android CSS hotfix is no longer needed, since we are no
longer pulling in the entire skins.minerva.base.styles module (including
the offending rule that broke app theming).

Bumps the CSS major version to reflect the major architectural change.

Bug: T220045
Change-Id: Ib91c1923ca235d952dd1f81a69dc52ef0a36aa44