CSS: Include external link icon CSS rule
Adds the rule adding an external link icon to our local CSS, for serving
through the /data/css/base endpoint.  The icon was still appearing in
the app when a theme was applied because it was being applied by via
wikimedia-page-library in ThemeTransform.css.

The rule is taken from
resources/src/mediawiki.skinning/content.externallinks.less in MediaWiki
core. There is a similar image provided as a ResourceLoader module in
the Minerva skin, but the implementation there appears
ResourceLoader-specific.

Bug: T219411
Change-Id: I2df020f228e1e642159ac65ada689a1c34ae9a96