lib/css.js: Use the path relative to lib/ to get the CSS file
When MCS starts, service-runner sets the CWD to the repo's root
directory, so relative paths will be relative to that path. However,
things change between source and deploy repos: to get to
`private/base.css` in the dpeloy repo, we need the relative path
`src/private/base.css`. In order to be able to satisfy all cases, make
the file path relative to the directory of `lib/css.js` rather than the
process' current CWD.

Change-Id: If6f6bfa43ecf6b07e370e93c132d22e652324b25