Update dependencies
Updates all dependencies except the eslint config one, which will come
in a later commit, since that one requires a few other files to change.

The main trigger for this was the update of the service-runner library,
which gets updated by the deploy prep process, and the desire to avoid
having a dirty repo after a deploy prep.

Other major (prod) dependency changes:
* core-js from 2.5.1 to 3.0.1

One dev dependency coulnd't be updated all the way since it seems to
require Node > 6:
* csv-parse: 4.4.0 (only updated to 3.2.0 instead)

Change-Id: If4d38252f3da4e095ee6ab3ae842c5f93ae0dc52