Dev: Structure MinervaNeue LESS files copies like original
To make future update easier let's keep the structure similar to the
original files as much as possible.

The diff in the base.css file is a result of moving the tablet styles
after the content styles, as is done in the original repo.

Also started documenting the update process in the css.md file.

Change-Id: I4278328380d6b739e0f72692577816fc49f5e382