Dev: use local CSS and JS when running locally
The config key mobile_html_rest_api_base_uri can be used to override the
value for the base URI to be used in the mobile-html head element.

mobile-html passes the mobile_html_rest_api_base_uri value as baseURI in
an options object to the processing script that needs it.

Change-Id: If9f243c633bdaf8c7847caa2670a085ec87a0a6e