Media: Filter out items missing imageinfo when building results
The action API imageinfo request will not return any imageinfo property
in the result for a given File page if the page has been deleted; and
the page media endpoint code is not prepared to handle this situation
(as seen in T223989). This updates the code to filter out items missing
any imageinfo.

Bug: T223989
Change-Id: I802c61b20dc62fe757aba86823b4f1a36054200e