Media: Improve codec parsing from type attributes
The existing codec parsing code in lib/media.js is parsimonious but too
brittle.  It expects codecs to be listed as a segment of the type
attribute of media source tags, and crashes if that segment doesn't
exist. For example, this is currently happening for
http://localhost:6927/it.wikipedia.org/v1/page/media/Barack_Obama.

This splits codec parsing into a more resilient standalone function,
and adds unit tests for it.

Side fix: whitespace inconsistencies in
test/lib/media/media-test-metadata.js.

Bug: T223989
Change-Id: I1f9f55ab47b807d5dd16d0e2f5186f8ead8ef096