Filter out false positives when querying page elements for media items
The media item selectors used when querying the page HTML for media
items are broad (for example, '*[typeof^=mw:Image]'; see the
MediaSelectors constant in  lib/selectors.js for the list).

Occasionally, this results in an invalid candidate, which is not
structured as an expected media item according to the HTML spec. These
cases should be filtered out as part of the filtering callback before
attempting to gather further metadata.

This is currently causing errors on:

http://localhost:6927/ar.wikipedia.org/v1/page/media/%D9%85%D8%AD%D9%85%D8%AF

Bug: T223989
Change-Id: I582f54f2fec60aae141f0d3325826fc26c6b4be3