Summary: Strip additional classes in preprocessing
The current revision (891354485) of the article "Financial statement"
on enwiki is breaking lead introduction parsing because it leads with
a misplaced citation and fixup template. These are not breaking extracts
generated by TextExtracts because it strips content with the selectors
sup.reference and .noprint. We should do likewise here (with
sup.reference translated to its Parsoid version, sup.mw-ref).

Bug: T225474
Change-Id: I13ee10c574d73c940938261582174b8a9e46f06c