references: stop modifying ids
Previously we removed the `cite_note-` prefix from ids. This lead to
confusion for clients and didn't provide enough benefit to make it
worthwile.

Interestingly this also has the nice side effect of fixing the order
of references in the list. See changes in the diff test results of the
Neutronium page.

Bumping major version since it's compatibility breaking.

Bug: T228476
Change-Id: I67d152f66b72db32f1b0f2250ff437005ca593a7