Fix: Metadata: Bail out of langlinks if titles are empty
There is a magic word, noexternallanglinks, that empties out the titles
of langlinks. In the web UI this causes them to redirect to the main
page.  That possibility wasn't handled here, leading to an error in
mediawiki-title.

This adds a check and bails out if an empty langlink title is
encountered.

Bug: T228936
Change-Id: I8ec2f9341d47b6bd3c2e419be2d8ee7f253dbd06