Performance: Remove language variant info from page metadata response
Removing variant info from the metadata endpoint dramatically improves
its performance. This is likely because it involves a number of slow MW
API calls. Variants are now being handled directly in the main page content
endpoints and so this info isn't needed.

The endpoint is currently marked experimental, and in practice no known
clients are relying on it, so it is safe to remove.

Bug: T229286
Change-Id: I229256d0f304d2ae36e9f0b3efd3b9826b3efb28