Fix new createDocument param type check
Recently added type checking for the HTML string passed in to
createDocument was mistakenly rejecting empty strings (''), since these
are falsy.

The second branch was also failing for empty strings, which have no item
at index 0.

This simplifies the check to just (typeof html === 'string'), which is
what we want.

Bug: T229849
Change-Id: Ifd4b00f81910a5eb890a83ba4e4e193f08c77f9a