Hygiene: Update media diff fixture
The diff tests are occasionally failing because of an occasional
inconsistency in the precision with which a video file's length is
reported in the /page/media endpoint. The inconsistency is because some
appservers are running PHP 7.2 while others are still running HHVM. See
T230181 for discussion. The failures are in fact becoming more frequent
as appservers continue to move to PHP 7.2.

Here the page currently used is swapped out for a different (smaller)
page which will hopefully not exhibit the same issue.

Change-Id: I3faa39b643c40dc5f256b61f58a64a54db399c2f