Dev: change spec example page for the talk endpoint
to an actual Talk page.

Redo of I2ff425b650212ee9c23aad3017234f423c898283, which got lost after
reapplying the patch to switch to OpenAPI 3.

The old page is too big and causes issues with the endpoint checker
timing out often, which is especially painful when deploying to prod.

Change-Id: If659652caa09328315fb6e18ba32ba8e9c77b2f4