Fix: Use Node.appendChild in addPageHeader when no first section exists
Fixes an error thrown in Domino (which was picked up by a diff test, but
not any regular unit or integration test) when attempting to insert the
header before a nonexistent first section in addPageHeader. A unit test
is added for the case.

Note: for no obvious reason, this was showing up as a timeout in
affected pages rather than a thrown exception.

Change-Id: I852d5a078836f56011e5a384cddef984f643a819