Tweak wikiLanguage.getLanguageCode to handle all needed domains
The wikiLanguage.getLanguageCode function currently only handles domains
of the form {lang}.wikipedia.org. This is a problem because we also need
to handle at least Beta Wikikipedias (e.g.,
es.wikipedia.beta.wmflabs.org) and the Wiktionaries (ar.wiktionary.org).

The update proposed here removes the check that domains contain
'wikipedia' as the second-to-last domain element. The first subdomain
segment will be returned for any domain with three or more parts, unless
the result is 'www', in which case we'll continue to return undefined.
This handles all of the cases we need to handle.

This means that the function will now attempt to parse a language code
from arbitrary domains, but that's should not block the change, since
we'll only encounter Wikimedia-controlled domains when running in
production.

Change-Id: I131a9700df42644e212543066595c153b4a6dad1