Featured-image: Stop limiting error handling for aggregated requests
Several feed endpoints have graceful error handling for aggregated
requests, meant to allow RESTBase to gracefully omit featured feed
content where the underlying content is not found as expected.

For the featured image endpoint, this is limited to 504 responses for
some reason. That doesn't seem necessary or desirable, so this removes
it.

Also removed some unused imports.

Change-Id: Id5282325b51ae2eb3514a2ee56afbb7c755c4aff