Media: Fix TypeError when processing pages with only Mathoid images
Fix a recent change to an imageinfo.makeResults call to pass `meta &&
meta.pages` where `meta` is a MediaWiki API result. Where no pages exist
in the result (as, for example, when a page contains only
Mathoid-generated math formula images), `meta.pages` may be undefined,
causing an error when .filter() is subsequently called on it.

Bug: T235408
Change-Id: Ida82a5980136d9c9e6b48749e404deaec6d2fd30