Speed up media endpoints, prevent them from blocking the event loop
- Use mutil to build the parsoid document instead of parsing synchronously
- Filter elements while walking the DOM

I tested with enwiki United_States/921631016. Before this change, the event loop would be blocked for ~200-300ms on my machine, after the change it's only blocked for ~20-30ms. This could be taken one step further to match the approach in https://gerrit.wikimedia.org/r/c/mediawiki/services/mobileapps/+/544237 and take a break after a certain number of nodes or time have passed.

Bug: T229286
Change-Id: Ie1524560710a1df1683c8ecfec93d1c3f9f3b8c0