Bring wikimedia-page-library into the mobileapps repo
Bring pagelib js and css into the repo, create new pcs css and js endpoints while keeping the legacy pagelib endpoints for existing app clients to allow for breaking changes

The wikimedia-page-library package is still in package.json because the legacy transforms rely on it. The package will be removed by the patch that introduces the mobile-html changes that removes these transforms.

Bug: T235571
Change-Id: I6d26a81b16777fc8c52191253b090b216b5fef17