Fix: Apply language headers to mobileview (zhwiki) mobile-html response
The mobile-html responses for zhwiki are currently not receiving
Content-Language headers, and do not have Accept-Language in their Vary
headers. This means that clients may be inappropriately falling back on
cached responses, particularly in the case where a user switches from
zh-hans to zh-hant. This patch updates the code to apply the expected
language headers for zhwiki responses.

Bug: T238916
Bug: T239659
Change-Id: I3350b82f66d090e8aeb3d5fcc337c76664f4ee8a