Mobile-html: Catch and rethrow URIErrors as HTTPErrors
URIErrors are occasionally being thrown when calling decodeURIComponent
while attempting to get the Parsoid link title in the course of creating a
MobileHTML object. This catches and rethrows these as HTTPErrors so that
more info on the request context (for example, the request URI) is
logged. This will better allow us to understand the underlying cause.

Bug: T240274
Change-Id: I81a7b672ea1e042dc5aa11a8bd0e2d41dbafb912