Dev: Change invalid section logging level from warning to debug
Currently a warning is logged when the service encounters a section that
appears to be malformed. Many of these are logged each day,
representing by far the largest category of mobileapps messages in
Logstash; yet there is no apparent issue in practice, and nobody is
following up on the warnings.

Given the above, the logging level of warning is too high. It should be
changed to debug.

Change-Id: I1759aa2e46798e99f695d6f2e5c262aace63f4d2