Mobile-html: URL-encode isVersionOf href value for mobileview (zhwiki)
In Parsoid HTML, the page URL provided in the 'href' properties of the
link[rel="dc:isVersionOf"] element is percent-encoded, and this is what
domUtil.getParsoidLinkTitle expects when calling decodeURIComponent on
it. However the URL is not currently encoded when adding
Parsoid-equivalent headers to mobileview HTML for zhwiki support, which
causes a URIError when certain values are provided - page titles
containing the '%' character, for example. This patch adds a call to
encodeURIComponent so that the value is in the expected format.

Bug: T240274
Change-Id: I9de55cb2f6778a528badd24edb45a7ce75715df5