pagelib: eslint rules
Softening some rules to reflect reality of past code changes.

(1) Increase max-len to 200
There is a lot of code now that has lines longer than 100 character.
To reduce the eslint noise double the value from 100 (from
eslint-config-node-services) to 200 for now.

(2) Effectively disabling no-use-before-define
I would have preferred to just disable it for arrow functions but that
doesn't seems to be an option yet[1].

[1] https://github.com/eslint/eslint/issues/10589

Change-Id: I4f9481b9f909f38febc66c5843565d4bb624feaa