Fix errors blocking zh mobile-html requests
The logs showed some `fatal/service-runner/unhandled` errors.[1][2]
Unfortunately, they don't show the full request URLs so it's difficult
to reproduce these errors.
I suspect that the issue fixed in I990af60d8ab7 was also a request to
zhwiki mobile-html and with this patch applied it would reveal the
offending URIs in the respective log entries.

There was a missing return of the promise to get the mobile-html from
action=mobileview. This caused the response to be blocked for a long
time when an error, such as a missing title was encountered.

After this patch the response comes back quickly and the log entry
includes the request information.
QA example:
http://localhost:8888/zh.wikipedia.org/v1/page/mobile-html/x12345dfg
                                             missing title ^

[1] 404: missingtitle
  https://logstash.wikimedia.org/app/kibana#/doc/logstash-*/logstash-2020.03.26/mobileapps?id=AXEYXFUFh3Uj6x1z7zzh&_g=h@44136fa
[2] TypeError: Cannot read property 'replaceChild' of null
  https://logstash.wikimedia.org/app/kibana#/doc/logstash-*/logstash-2020.03.26/mobileapps?id=AXEYnjycarkxubcmN7FF&_g=h@44136fa

Change-Id: Iad3ae0181dea7c44b6240f6633e986dda3bd0d81