Revert "Fix possible TypeError when preparing red links"
This reverts commit ccbc3d8bc3453e2c3013794b7bf4118edbe129f1.

Reason for revert: I would like to see which pages are causing this issue. Now that we have Iad3ae0181dea7 we could let this run for a few days and then bring back this fix if needed. It could mask other underlying issues since I think the only we run into this is when the parentNode is falsey. (The element is already checked in the calling function.)

Change-Id: I616950ab48985f77369b6ca83bfa7612e24985e0