Hygiene/Dev: Use hostname of the request for base API URIs
Allows for a single conf to allow for requests to localhost, 127.0.0.1, or your local IP. Useful for testing with the Android app's new localPCSDebug build variant that points to the service running locally, but needs to use Android's magic 10.0.2.2 IP address due to the emulator not sharing a network interface with the host machine.

Also adds wikipedia.org and *.wikipedia.org to the img-src. Some on-wiki images are still hosted on wikipedia.org instead of wikimedia.org. See enwiki > Hathor for an example failing request.

Change-Id: I18f607cac036136e3fd848f0a3fb0c348040553e