Add footer if query parameter is used
If query parameter "footer=true" is added to the mobile-html URL
then the footer is automatically added.
Can be combined with the "theme" query parameter.

The footer works better when going via RESTBase, so the read more
results can be included.

For local environments local changes to config recommended:
  # use_request_host: true
  mobile_html_rest_api_base_uri: //localhost:7231/meta.wikimedia.org/v1/
  mobile_html_local_rest_api_base_uri_template: //localhost:7231/{{domain}}/v1/

Examples:
http://localhost:7231/en.wikipedia.org/v1/page/mobile-html/A?footer=true&theme=sepia

Or once this is merged and deployed:
https://en.wikipedia.org/api/rest_v1/page/mobile-html/A?footer=true&theme=dark

Change-Id: Ie216855845d7567df4ca8091fd0bdbc163c0cc9d