Update edit section preview to have identical iOS app output as pre-PCS.
PCS treats edit previews differently than old system. When previewing an edit section, there shouldn't be an article header, sections should not be collapsible, and the references section should have a header of "Preview of references". Additionally, moved logic to hide pencil icons for edit previews to this same file.

Bug: T251473
Change-Id: Ifa7f0b69119c57dc3021c76962b0dc8b78cdb22d