Hygiene: update diff test results
Parsoid now adds an nbsp inside a span like this[1]:

double talk page bankrupt!
<span typeof="mw:DisplaySpace mw:Placeholder" id="mwKg">&nbsp;</span>
;&gt; Just

Not 100% how it was before but after our transformations this
resulted in a single space character instead.
Now it's the nbsp (%C2%A0).

Timing: this diff test broke shortly[2] after the MW deployment train
finished, so it stands to reason that this was likely the result of a
(PHP-)Parsoid change.

[1] https://en.wikipedia.org/api/rest_v1/page/html/User_talk:Brion_VIBBER/895522398
[2] https://integration.wikimedia.org/ci/job/mobileapps-periodic-test/18200/console

Change-Id: If76c65906e63ce627dd66b82027fa2254dddff91