Only pass the first relevant language code along to action=mobileview
It appears that mobileview only handles a single language passed through
the Accept-Language header. Ensure that if the user sends multiple
languages to Accept-Language, only the first relevant one is passed to
action=mobileview.

Updated the diff test for the currently used Chinese article to test
mobile-html output to also get a list of languages in the
accept-language header.

Bug: T254627
Change-Id: I44461b2460f7dd95a7659f1dd2fdb5f42f01832e