Disable touch callout on all img tags, make image src schemeless
Follow on-fix for images that aren't a timeline. The touch menu still works with gallery images after this fix. This also addresses an issue where math images wouldn't load locally because they are hardcoded to the https:// scheme and https://wikimedia.org isn't allowed by the CSP. They should be schemeless like other URLs on the page to work with the app:// scheme on iOS.

Bug: T254781
Change-Id: I310cab0c5c7a35dec5786c2b3db9ed2f2f019305