media-list: include timeline images
Before this timeline images were not included in the media-list putput.
The goal of this patch is to fix that.

Timeline images have a different typeof attribute value in Parsoid from
regular images, namely `mw:Extension/timeline`.
This patch introduces a new internal type called, TimelineImage.
(Externally, the type is image.)
This patch is currently following the scheme used for MathImage items to
set an original property in the result. Alternatively, we could consider
guarding against original being falsy when removing duplicates.

This currently assumes that the timeline image is a PNG. I think they
all are but I'm not sure. If it's not a PNG then we skip it for now.
Actually, I'm not sure why we need to set a mime type here. It doesn't
seem to be used in media-list.

Bug: T254290
Change-Id: I21097fcd4d4f23b8f2c30c9f26eaf8829c634b84