Footer.add: remove title param
Clients pass the page title but sometimes the title needs to be
percent-encoded.

This patch derives the page title from window.location instead of accepting what
the clients provide. (The client title is ignored.)
If we want to go that route clients lose the ability to get the list of
related pages for another title but that seems only be useful for demo
purposes.

Testing this locally:
Using `?footer=true` does still not work since `baseURL` also needs to
be overridden to point to a RB instance.
Instead you can add the footer manually:
```
pcs.c1.Footer.add({
  platform: 'ios',
  version: '2',
  menu: {
    items: [pcs.c1.Footer.MenuItemType.lastEdited, pcs.c1.Footer.MenuItemType.pageIssues, pcs.c1.Footer.MenuItemType.disambiguation, pcs.c1.Footer.MenuItemType.talkPage],
    fragment: "pcs-menu",
    editedDaysAgo: 3
  },
  readMore: {
    itemCount: 3,
    baseURL: 'https://en.wikipedia.org/api/rest_v1',
    fragment: "pcs-read-more"
  }
})
```

Bug: T255579
Change-Id: I0481e77c646163a1a415ff7f4ca33d8d1b6aa1ac