Site-CSS: Fix load.php request to use request templates
The switchover to Kubernetes exposed a bug in the PCS site CSS endpoint
in which raw preq requests were being made to the public load.php
endpoints for CSS styles from ResourceLoader. These requests should be
made to cluster-internal appserver endpoints, but succeeded anyway when
made from the legacy Service Cluster B deployment. The requests fail
when made from the Kubernetes cluster.

This change updates this request to use request templating so that an
internal MediaWiki appserver endpoint can be defined in the production
configuration (i.e., in operations/deployment-charts).

Bug: T258186
Change-Id: I21cdca53cc9312e742f38c04d6718125e9165271