Make Spoken Wikipedia filename parsing more robust
Spoken Wikipedia filename parsing currently relies on Parsoid's
JSON representation of a template, as provided in the HTML as a
`data-mw` attribute on the leading element. This solution is brittle,
and was indeed broken by a recent change to
`Template:Spoken_Wikipedia` on English Wikipedia, causing one of our
unit tests to currently fail.

The Spoken Wikipedia filename parsing function is updated here to rely
on the page HTML structure itself rather than the `data-mw` attribute
for the template.

Change-Id: Ifc681e6b536b0be1bd9c3510c914b8512fefb49b