Hygiene: Remove unnecessary pagelib diff tests
Removes the diff tests for the pagelib CSS and JS endpoints. These
are unnecessary because the diff tests are meant to catch upstream
changes that could affect our service's output, such as changes to
Parsoid or a wiki's global site CSS or JS. The pagelib inputs
consist solely of source code within this repo, so there is nothing
for the diff tests to guard against.

Change-Id: Ie3bd9a47543a9e822a0c77fdbfe40770d44e49b4