mobile-html: fix some hatnotes not being hoisted properly
Even though the templates add the hatnote class to a DOM element
it's not properly hoisted. The reason is because we skip when we don't
see a hatnote class in the DOM element, but in this particular case
the transform is skipping on a style tag, that should be skipped anyway
if it's the first element.

Bug: T277465
Change-Id: I939ba9828bab6ce8e35ff5d736f31c45648e3647