Allow isEditable to be an initial setup parameter.
We're introducing a feature that will optionally show or hide the
edit pencil(s) on articles during the user's browsing session.
Currently the page library offers a function to set the visibility
of the edit pencils, but only after the page has already loaded,
which means the edit pencil will be momentarily shown before being
hidden. It would be smoother to allow the pencil to be hidden or
shown in the initial setup sequence.

Bug: T254771
Change-Id: Ib34e1897c34f1def3f2e0626008446ce43433d99