Correctly scroll to top of collapsible tables when collapsing from bottom.
When a table (e.g. infobox) is expanded, it has a "Close" footer
at the bottom. When the footer is tapped, the table collapses, but
doesn't scroll back up to the top of the table, leaving the user
in an unexpected point in the page.

This will automatically ensure that when the table footer is
tapped, the page will be scrolled up so that the header of the
table is in the center of the screen.

Bug: T301075
Change-Id: I8f48962d02059a3edde6298978cb3d44bdc823f7