dirtyfix mobile zebra table dark mode issue
This addresses the long-known issue of zebra tables not being readable
in mobile dark view because the text and the background are white on white
on every other table row. This is not an ideal fix, but there seems to be
no attempt made yet to find a more thorough, clean fix. This annoys
thousands of users every day (the dark mode is practically not usable
if you are reading articles including striped/zebra tables often), so
a quick fix is greatly appreciated. See also the dozens of bug reports
regarding this topic.

Bug: T264474
Change-Id: I7baddb3d5ab80b7f6e65a36c8b0dccf9697dcbfd