Revert "Fix failing CI"
This reverts commit cee4e5935a220769c8c29d6a8d9cf6093fb5e322.

Reason for revert: No longer necessary.
https://test2.wikipedia.org/wiki/FlaggedRevsTest has been moved back,
and will hopefully remain there. https://test2.wikipedia.org/wiki/Test,
on the other hand, is not designated to be used for CI, so it can break
any time. And it doesn’t even test what this test case is ought to
test: the test case should test whether the code properly handles
unreviewed pages, but Test is reviewed (it only has some pending template changes).

Change-Id: I0c191f1a075b28d6b667b49e39311df9faeb4eed