Fix setting of background color for Android dark/black themes.
In the previous patch (906699) for setting the background color, the
color was defined in Constants.less as a "@"constant, but then
referenced as a var() variable, which is not correct. The result is an
undefined or "transparent" background color.

The reason this "worked" when testing in Android is because the app
itself provides a background color that matches the background
defined here, so the transparent background of the WebView wasn't
apparent.

This patch corrects the usage of the background constant, so that it's
applied correctly in the output CSS.

Bug: T332823
Bug: T335035
Change-Id: I02b4b12d5bcb03836183308f2747b3903abb4aab