No longer use page/related endpoint, and offer lazy loading.
This is intended to be backwards-compatible with existing (and previous)
versions of apps in the way they call the Footer.add() function, i.e.
no changes will be necessary on the apps' side.

However, this also introduces a way for the client to lazy-load the
Read More items, by passing a parameter of "readMoreLazy=true". If this
parameter is given, then the Footer.add() function will skip adding the
Read More items, and the client will need to call Footer.appendReadMore()
when it's ready to actually load the items. The parameters for
Footer.appendReadMore() are the same as for Footer.add().

The API query used here is literally identical to the query used in
mobile web, so that both of these queries can share cache.

Add test for showReadMore transformation

Bug: T352801
Bug: T353188
Change-Id: I7d26d686e98d5d534dcaee6d859fc318af0dca4e