Fix horizontal scrolling regression
The regression was introduced by
I56f8003a56eedaa0485f7fc4cd7d8b4cd94ffad0
which effectively reverted the previous fix in
I3a83ec2dc5d449c4fc7584e05ceb4a61ef61ffa9

Displaying the figure as a table accommodates the width of its contents,
which is by design, since the MediaWiki parsers place the width on the
media elements.

However, the issue doesn't exist when the media is loaded, only when the
placeholder is present with its inline styles.  The fix is to let the
pcs-widen-image-override class styles apply to the placeholder span as
well as the img.

@TheDJ pretty much explained the issue in T312667#8087809

Bug: T312667
Change-Id: I7274be2d8233addef4e33419b6cc7a04c020cef6