Fix build of CSS, due to new math rules.
The latest version of Less has updated treatment of math
symbols, especially the division operator which is used
throughout CSS in other ways.

The default "math mode" in Less now requires the division
operator to be surroundeed by parentheses. But for compatibility
purposes, we can specify a parameter of "math=always" and
make it treat the division symbol the same as before.

(This usage of math is deprecated, and will eventually need
to be updated with proper parentheses upstream in Minerva.)

https://lesscss.org/usage/#less-options-math

Change-Id: Ia7143ff47c412480c351b789245fa787ada71d71