Fix section/heading formatting for non-parsoid requests.
This should fix the parsing and manipulation logic that is causing
table-of-contents functionality to fail for Chinese articles.

The root cause seems to be that the mobile-html endpoint generally
uses Parsoid for the article contents, *except* in the case of
Chinese, whereby it falls back to using action=parse. Previously
it used action=mobileview, but this was deprecated and replaced
with action=parse. It is possible that mobileview was producing a
specific structure that our parsing logic was expecting, which
action=parse no longer produces? (yet still passes tests?)

Bug: T370569
Change-Id: I5ee15b922e8db443adbf0d15425ee9c7ed1fb1dc