Update and modernize image lazy-loading, take 2.
This removes most of the old logic for lazy-loading images,
and instead uses the modern loading=lazy attribute.

We are actually keeping the server-side logic of transforming
img tags to spans, and the client-side logic of transforming
spans back to imgs, except the client-side logic will run all
at once, instead of lazily. This has a couple of advantages:

* It will not require cache invalidation of existing articles,
i.e. the updated logic is "compatible" with the current HTML of
articles that are cached.
* The apps have a client-side preference of not loading images
at all, which this will continue to respect, since it relies on
the same logic. (If image loading is disabled, the spans will
simply never be transformed into imgs.)
* This will improve efficiency, since we will no longer monitor
DOM scrolling, and allow the browser to lazy-load images natively.

Note: This also includes an update to the blubber file to fix
testing in CI, which requires writing to the filesystem.

Bug: T328943
Change-Id: I98e477c004408cb8d9dc7f57ba01284d4e399c75